requirements.txt
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
click==8.1.3
colorama==0.4.3
comm==0.1.2
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.6.5
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
future==0.18.2
greenlet==2.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.4.9
jupyter-console==6.4.4
jupyter-core==5.1.3
jupyter-events==0.6.3
jupyter-server==2.1.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.5
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.10.4
notebook==6.5.2
notebook-shim==0.2.2
oauthlib==3.1.0
olefile==0.46
packaging==23.0
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pkgutil-resolve-name==1.3.10
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.6.1
psutil==5.9.4
psycopg2==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.14.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
python-json-logger==2.0.4
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
stack-data==0.6.2
systemd-python==234
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.8.1
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.17.1
wadllib==1.3.3
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
xkit==0.0.0
zipp==3.11.0
